--- src/main/java/com/tech/Models/ImagesDAO.java
@@ -1,18 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package com.tech.Models;
-
-import java.util.List;
-
-/**
- *
- * @author KuroiTenshi
- */
-public interface ImagesDAO {
-    public void save(ImagesDAO imagesDAO);
-    public List<ImagesDAO> list();
-    
-}
