+++ src/main/java/com/tech/Models/daos/interfaces/ImagesDAO.java
@@ -1,0 +1,18 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package com.tech.models.daos.interfaces;
+
+import java.util.List;
+
+/**
+ *
+ * @author KuroiTenshi
+ */
+public interface ImagesDAO {
+    public void save(ImagesDAO imagesDAO);
+    public List<ImagesDAO> list();
+    
+}
