+++ src/main/java/com/tech/services/ImagesService.java
@@ -5,9 +5,11 @@
  */
 package com.tech.services;
 
-import com.tech.Models.ImagesMod;
-import com.tech.Repositories.IImagesRepository;
+import com.tech.models.entities.ImagesMod;
+import com.tech.repositories.IImagesRepository;
 import java.util.List;
+
+import com.tech.services.interfaces.IImagesService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -17,7 +19,7 @@ import org.springframework.transaction.annotation.Transactional;
  * @author KuroiTenshi
  */
 @Service
-public class ImagesService implements IImagesService{
+public class ImagesService implements IImagesService {
     
     @Autowired
     private IImagesRepository repository;
