+++ src/main/webapp/WEB-INF/rest-servlet.xml
@@ -10,7 +10,7 @@
         http://www.springframework.org/schema/context/spring-context-4.0.xsd
         http://www.springframework.org/schema/mvc
         http://www.springframework.org/schema/mvc/spring-mvc-4.0.xsd">
-        <context:component-scan base-package="com.tech.Controllers" />
+        <context:component-scan base-package="com.tech.controllers" />
         
         <mvc:annotation-driven />
