+++ src/main/webapp/images.html
@@ -16,7 +16,7 @@ and open the template in the editor.
          <div id="body-wrapper">
 <!--             <table>
                 <tr>
-                    
+
                     <td>images: </td>
                     <td>
                         <input type="file" name="file" id="images">
@@ -32,9 +32,8 @@ and open the template in the editor.
              
            <form method="POST" enctype="multipart/form-data"
 		action="/upload/c">
-		File to upload: <input type="file" name="file"><br /> Name: <input
-			type="text" name="name"><br /> <br /> <input type="submit"
-			value="Upload"> Press here to upload the file!
+		File to upload: <input type="file" name="file"><br /> Name: <input type="text" name="name"><br /> <br />
+               <input type="submit" value="Upload"> Press here to upload the file!
             </form>
             <br /><br />
