+++ src/main/webapp/index.html
@@ -17,6 +17,9 @@
                     <td>
                         <a href="/chat.html">Chat</a>
                     </td>
+                    <td>
+                        <a href="/images.html">Images</a>
+                    </td>
                 </tr>
             </table>
         </div>
